feat: Add route to retrieve statistics
The new route `/statistics` returns statistics about the data stored in the database, including the total number of notes, the total number of correct notes, and the percentage of correct notes. This allows users to get an overview of the collected data.

A `get_statistics()` function has been added to the `app.py` file that performs the necessary queries to the database and returns the results in JSON format.

Additionally, the `index.html` file has been updated to display the statistics on the web page. HTML elements have been added, and an `updateStatistics()` function has been created that makes a request to the new `/statistics` route and updates the values on the page.

The frequency vs time graph has also been updated to show the last 10 data points instead of all the collected data. This has been achieved using two arrays, `timeData` and `frequencyData`, which are updated each time new data is obtained. The `createGraph()` function has been modified to initialize the graph with these arrays, and the `updateData()` function has been updated to add the new data points to the arrays and update the graph.

These modifications allow users to efficiently visualize the statistics and graph with a better user experience.